requirements.txt
# Superstore Document Pipeline Requirements
pytest>=7.0.0
pyyaml>=6.0
